% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{wa_pal}
\alias{wa_pal}
\title{A Collection of Western Australia Colour Palettes}
\format{
An object of class \code{list} of length 1.
}
\source{
\describe{
\item{Lake Hillier}{\url{https://upload.wikimedia.org/wikipedia/commons/f/fb/Lake_Hillier_Shoreline_Pink_Hue_Salt_Deposite.jpg}}
}
}
\usage{
wa_pal
}
\description{
A collection of colour palettes inspired by the landscape and wildlife of
Western Australia. The list of available palettes is:
\describe{
\item{Lake Hillier}{Diverging palette based on the colours of Lake Hillier}.
}
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

library(ggplot2)
df <- data.frame(
  x = rnorm(100, -20, 20),
  y = rnorm(100, -20, 20),
  cl = sample(letters[1:8], 100, replace = TRUE)
)
ggplot(df, aes(x, y, colour = cl, shape = cl)) +
  geom_point(size = 4) +
  scale_colour_wa("Lake Hillier") +
  theme_bw() +
  theme(aspect.ratio = 1)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
